+++ pymc3/tests/test_sampling.py
@@ -103,7 +103,7 @@ class TestSample(SeededTest):
             assert "'foo'" in str(excinfo.value)
 
             with pytest.raises(ValueError) as excinfo:
-                pm.sample(50, tune=0, init=None, step_kwargs={"foo": {}})
+                pm.sample(50, tune=0, init=None, foo={})
             assert "foo" in str(excinfo.value)
 
             with pytest.raises(ValueError) as excinfo: